chore: relocate go.mod to repo root, rename module to leftmark
Flattens src/ into the repo root so go.mod, internal/, and main.go
sit together without a stuttering import path, and renames the
module from the stale notesmd.com/cli to github.com/issam-assiyadi/leftmark.
First step of the domain-driven restructure: makes the domain/
application/adapter packages plannned next importable outside
internal/ without churn. Also fixes the golangci-lint go version
(1.25 -> 1.23) to match go.mod, and clears 6 pre-existing errcheck
lint failures so `make check` passes cleanly going forward.
