+++ .golangci.yml
@@ -1,7 +1,7 @@
 version: "2"
 
 run:
-  go: "1.25"
+  go: "1.23"
 
 linters:
   default: none
