+++ scripts/vet.sh
@@ -5,5 +5,4 @@ set -eu
 go_cache="$1"
 
 mkdir -p "$go_cache"
-cd src
 GOCACHE="$go_cache" go vet ./...
