--- src/Makefile
@@ -1,16 +1,0 @@
-.PHONY: fmt fmt-check vet lint check
-
-fmt:
-	$(MAKE) -C .. fmt
-
-fmt-check:
-	$(MAKE) -C .. fmt-check
-
-vet:
-	$(MAKE) -C .. vet
-
-lint:
-	$(MAKE) -C .. lint
-
-check:
-	$(MAKE) -C .. check
